+++ src/write_outputs/dftranspose.jl
@@ -1,26 +1,37 @@
 ################################################################################
-## function dftranspose(df)
-##
-## inputs: df - [DataFrame] a DataFrame object to be transposed
-## results t - [DataFrame] a transposed version of the df DataFrame.
-##		   withhead - [Boolean] if True, first column of df will become column
-##		   names for t. Otherwise, first column will first row and column names
-##		   will be generic (e.g. x1:xN)
-##
-## Note this function is necessary because no stock function to transpose
-## DataFrames appears to exist.
+## helpers for writing transposed output tables
 ################################################################################
 @doc raw"""
-   df = dftranspose(df::DataFrame, withhead::Bool)
+    transpose_output_dataframe(df::DataFrame; withhead::Bool=false)
 
-Returns a transpose of a Dataframe.
+Return a transposed copy of an output DataFrame without collecting each row into
+an intermediate vector.
 """
-function dftranspose(df::DataFrame, withhead::Bool)
-    if withhead
-        colnames = cat(:Row, Symbol.(df[!, 1]), dims = 1)
-        return DataFrame([[names(df)]; collect.(eachrow(df))], colnames)
-    else
-        return DataFrame([[names(df)]; collect.(eachrow(df))],
-            [:Row; Symbol.("x", axes(df, 1))])
+function transpose_output_dataframe(df::DataFrame; withhead::Bool = false)
+    row_count, col_count = size(df)
+    colnames = withhead ? Symbol[:Row; Symbol.(df[!, 1])] :
+               Symbol[:Row; Symbol.("x", 1:row_count)]
+    transposed = Matrix{Any}(undef, col_count, row_count + 1)
+    headers = names(df)
+
+    for col in 1:col_count
+        transposed[col, 1] = headers[col]
     end
-end # End dftranpose()
+
+    for row in 1:row_count
+        for col in 1:col_count
+            transposed[col, row + 1] = df[row, col]
+        end
+    end
+
+    return DataFrame(transposed, colnames)
+end
+
+function write_transposed_csv(filename::AbstractString, df::DataFrame; kwargs...)
+    options = Dict{Symbol, Any}(pairs(kwargs))
+    if haskey(options, :writeheader) && !haskey(options, :header)
+        options[:header] = options[:writeheader]
+        delete!(options, :writeheader)
+    end
+    return CSV.write(filename, transpose_output_dataframe(df); options...)
+end