+++ src/write_outputs/write_storagedual.jl
@@ -1,9 +1,16 @@
 @doc raw"""
-	write_storagedual(path::AbstractString, inputs::Dict, setup::Dict, EP::Model)
+    write_storagedual(path::AbstractString, inputs::Dict, setup::Dict, EP::Model,
+        cache::OutputCache = build_output_cache(EP, inputs, setup))
 
 Function for reporting dual of storage level (state of charge) balance of each resource in each time step.
+The optional `cache` argument allows callers to reuse extracted model outputs across
+multiple write functions to reduce memory allocations.
 """
-function write_storagedual(path::AbstractString, inputs::Dict, setup::Dict, EP::Model)
+function write_storagedual(path::AbstractString,
+    inputs::Dict,
+    setup::Dict,
+    EP::Model,
+    cache::OutputCache = build_output_cache(EP, inputs, setup))
     gen = inputs["RESOURCES"]
     zones = zone_id.(gen)
 
@@ -23,25 +30,25 @@ function write_storagedual(path::AbstractString, inputs::Dict, setup::Dict, EP::
 
     # # Dual of storage level (state of charge) balance of each resource in each time step
     dfStorageDual = DataFrame(Resource = inputs["RESOURCE_NAMES"], Zone = zones)
-    dual_values = zeros(G, T)
+    dual_values = resource_time_scratch!(cache)
 
     # Loop over W separately hours_per_subperiod
     if !isempty(STOR_ALL)
         STOR_ALL_NONLDS = setdiff(STOR_ALL, inputs["STOR_LONG_DURATION"])
         STOR_ALL_LDS = intersect(STOR_ALL, inputs["STOR_LONG_DURATION"])
-        dual_values[STOR_ALL, INTERIOR_SUBPERIODS] = (dual.(EP[:cSoCBalInterior][
+        dual_values[STOR_ALL, INTERIOR_SUBPERIODS] .= (dual.(EP[:cSoCBalInterior][
             INTERIOR_SUBPERIODS,
             STOR_ALL]).data ./ inputs["omega"][INTERIOR_SUBPERIODS])'
-        dual_values[STOR_ALL_NONLDS, START_SUBPERIODS] = (dual.(EP[:cSoCBalStart][
+        dual_values[STOR_ALL_NONLDS, START_SUBPERIODS] .= (dual.(EP[:cSoCBalStart][
             START_SUBPERIODS,
             STOR_ALL_NONLDS]).data ./ inputs["omega"][START_SUBPERIODS])'
         if !isempty(STOR_ALL_LDS)
             if inputs["REP_PERIOD"] > 1
-                dual_values[STOR_ALL_LDS, START_SUBPERIODS] = (dual.(EP[:cSoCBalLongDurationStorageStart][
+                dual_values[STOR_ALL_LDS, START_SUBPERIODS] .= (dual.(EP[:cSoCBalLongDurationStorageStart][
                     1:REP_PERIOD,
                     STOR_ALL_LDS]).data ./ inputs["omega"][START_SUBPERIODS])'
             else
-                dual_values[STOR_ALL_LDS, START_SUBPERIODS] = (dual.(EP[:cSoCBalStart][
+                dual_values[STOR_ALL_LDS, START_SUBPERIODS] .= (dual.(EP[:cSoCBalStart][
                     START_SUBPERIODS,
                     STOR_ALL_LDS]).data ./ inputs["omega"][START_SUBPERIODS])'
             end
@@ -49,19 +56,19 @@ function write_storagedual(path::AbstractString, inputs::Dict, setup::Dict, EP::
     end
 
     if !isempty(VRE_STOR)
-        dual_values[VS_STOR, INTERIOR_SUBPERIODS] = ((dual.(EP[:cSoCBalInterior_VRE_STOR][
+        dual_values[VS_STOR, INTERIOR_SUBPERIODS] .= ((dual.(EP[:cSoCBalInterior_VRE_STOR][
             VS_STOR,
             INTERIOR_SUBPERIODS]).data)' ./ inputs["omega"][INTERIOR_SUBPERIODS])'
-        dual_values[VS_NONLDS, START_SUBPERIODS] = ((dual.(EP[:cSoCBalStart_VRE_STOR][
+        dual_values[VS_NONLDS, START_SUBPERIODS] .= ((dual.(EP[:cSoCBalStart_VRE_STOR][
             VS_NONLDS,
             START_SUBPERIODS]).data)' ./ inputs["omega"][START_SUBPERIODS])'
         if !isempty(VS_LDS)
             if inputs["REP_PERIOD"] > 1
-                dual_values[VS_LDS, START_SUBPERIODS] = ((dual.(EP[:cVreStorSoCBalLongDurationStorageStart][
+                dual_values[VS_LDS, START_SUBPERIODS] .= ((dual.(EP[:cVreStorSoCBalLongDurationStorageStart][
                     VS_LDS,
                     1:REP_PERIOD]).data)' ./ inputs["omega"][START_SUBPERIODS])'
             else
-                dual_values[VS_LDS, START_SUBPERIODS] = ((dual.(EP[:cSoCBalStart_VRE_STOR][
+                dual_values[VS_LDS, START_SUBPERIODS] .= ((dual.(EP[:cSoCBalStart_VRE_STOR][
                     VS_LDS,
                     START_SUBPERIODS]).data)' ./ inputs["omega"][START_SUBPERIODS])'
             end
@@ -76,8 +83,8 @@ function write_storagedual(path::AbstractString, inputs::Dict, setup::Dict, EP::
     rename!(dfStorageDual,
         [Symbol("Resource"); Symbol("Zone"); [Symbol("t$t") for t in 1:T]])
 
-    CSV.write(joinpath(path, "storagebal_duals.csv"),
-        dftranspose(dfStorageDual, false),
+    write_transposed_csv(joinpath(path, "storagebal_duals.csv"),
+        dfStorageDual,
         header = false)
 
     if setup["OutputFullTimeSeries"] == 1 && setup["TimeDomainReduction"] == 1